buildbot: lint against base ref instead of sha
The current logic tries to lint against the merge-base of the PR and
GitHub's reported base sha. Unfortunately, the base sha isn't guaranteed
to exist yet on the buildbot, as it's only sometimes in the ancestry
graph of the PR.

The previous fix attempted to fetch the relevant sha directly from
GitHub before calculating the merge-base. This also failed, because
git isn't required to support directly fetching shas, only refnames.
(Confusingly, if you try this out locally, git will claim it
successfully fetched the sha if it already exists, rather than telling
you that that's an invalid operation.)

This PR gives up on passing the base sha and passes the base ref name.
For example, where before we would try to lint against "...918b037",
now we would simply lint against "...master".