central: Send PR status events before triggering builds
This was removed when I added support for Buildbot pending build events
as I had assumed that Buildbot would send build events fast enough.

However one problem that is made obvious by the occasional build upload
issues is that Buildbot only sends events when a build has started
and not when a build is triggered, which means that a builder getting
stuck can cause PRs to get the green "all checks passed" status
even when builds haven't actually started on some builders.