buildbot: switch to CMake+Ninja for Windows builds
Finishes the CMake migration.

Unlike #113 (the previous attempt), this PR takes a different approach.
Instead of assuming that the compiler and environment variables are
already set, we call the vcvars64.bat batch script to find the WinSDK,
compiler path, include paths, etc. and to set any environment variables
that are required to compile.

This should require less maintenance than the previous approach.
The only time we have to change something in the buildbot config file
is when we upgrade to a new major version of Visual Studio.

One concern that was raised last time is that the WinSDK version is not
pinned so we end up using the latest WinSDK version.
However, since https://github.com/dolphin-emu/dolphin/pull/8498 was
merged, the VS project files also do the exact same thing, and minimum
version checks have been added elsewhere to ensure the WinSDK version
isn't too old.

Anyway, with this approach, if we ever need to use a specific version
of the Windows SDK, we can easily switch to vcvarsall.bat later on:

https://docs.microsoft.com/en-us/cpp/build/building-on-the-command-line?view=msvc-160#vcvarsall-syntax

So I don't think this should be a blocker.